# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Admin API",
    "description": "Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "",
      "description": " Main server"
    }
  ],
  "paths": {
    "/admin": {
      "get": {
        "summary": "Checks if the server is alive",
        "parameters": [
          {
            "in": "header",
            "name": "x-auth-token",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully retrieved Admin details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/admin/login": {
        "post": {
          "summary": "admin login",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                   
                    "email": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                    
                      
                    }
                  }
                 
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Created"
            }
          }
        }
      
    ,
    "/admin/{id}": {
        "get": {
          "summary": "get admin detail",
          
          "responses": {
            "200": {
              "description": "successfully retrieved Admin details",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
    "/admin/change-password": {
      "post": {
        "summary": "change-password with old password",
        "parameters": [
          {
            "in": "header",
            "name": "x-auth-token",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "old-password": {
                    "type": "string"
                  },
                  "new-password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "password changed successfully",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/admin/reset-password-email": {
      "post": {
        "summary": "reset-password link to email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "password sent to email successfully",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/admin/reset-password/{id}/{token}": {
      "post": {
        "summary": "forgot password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "password changed successfully",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/admin/edit": {
      "put": {
        "summary": "update admin details",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "details edited successfully",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/admin/logout": {
        "post": {
          "summary": "logout",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": " admin got logout",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
  }

}